fix: export `DescriptionType` enum values
At compliation time the values of a `const enum` in TS get changed
to numbers which breaks JS APIs that expect the string values.

At runtime anything in a `.d.ts` file is not available, so the
solution is to export enum definitions from the `.d.ts` file as a
regular `enum` (no `const`) and export the matching values from the
implementation file as a plain object.