feat: accept preconfigured PeerConnection as polyfill arg
There are cases where we need to use methods on `PeerConnection` that
are not available on `RTCPeerConnection`, but then we wish to pass
the `RTCPeerConnection` on to other code for it to use in a polmorphic
node/browser way.

Add an option to the polyfilled `RTCPeerConnection` to pass a
preconfigured `PeerConnection` instance in, instead of always creating
one.