fix: return globalThis types from polyfill
This is being done in a few places so this PR just tidies up the
remaining bits.

Without this you can't assign a `RTCPeerConnection` (e.g. the polyfill
object) to a variable of type `globalThis.RTCPeerConnection` as
several fields are present in the polyfill but missing in the DOM
type.