test: recieve lots of data
Adds a test that opens a datachannel. When the remote receives notification of the new channel it writes lots of data into it and closes the channel.  The local peer counts the received bytes and asserts that it received them all.

This test currently fails intermittently, largely down to the amount of load the process is under at the time.

You can run similar code in a browser here: https://codepen.io/achingbrain/pen/KwVPZBP?editors=0012

This works in Chrome and Firefox Nightly, though Firefox is a bit slow.